+++ src/datatable.module.ts
@@ -36,7 +36,8 @@ import {
 
 import {
   ScrollbarHelper,
-  DimensionsHelper
+  DimensionsHelper,
+  ColumnChangesService
 } from './services';
 
 @NgModule({
@@ -45,7 +46,8 @@ import {
   ],
   providers: [
     ScrollbarHelper,
-    DimensionsHelper
+    DimensionsHelper,
+    ColumnChangesService
   ],
   declarations: [
     DataTableFooterTemplateDirective,